+++ README.md
@@ -5,9 +5,6 @@
 [![Code Climate](https://codeclimate.com/github/Swati4star/Images-to-PDF.svg)](https://codeclimate.com/github/Swati4star/Images-to-PDF) 
 [![Android Arsenal](https://img.shields.io/badge/Android%20Arsenal-PDF%20Converter-blue.svg?style=true)](https://android-arsenal.com/details/3/7132)
 
-
-[![Waffle.io - Columns and their card count](https://badge.waffle.io/Swati4star/Images-to-PDF.svg?columns=all)](https://waffle.io/Swati4star/Images-to-PDF)
-
 [<img src="https://f-droid.org/badge/get-it-on.png"
       alt="Get it on F-Droid"
       height="80">](https://f-droid.org/app/swati4star.createpdf)
