+++ app/src/main/java/swati4star/createpdf/fragment/SplitFilesFragment.java
@@ -38,8 +38,8 @@ import swati4star.createpdf.util.BottomSheetCallback;
 import swati4star.createpdf.util.BottomSheetUtils;
 import swati4star.createpdf.util.FileUtils;
 import swati4star.createpdf.util.MorphButtonUtility;
-import swati4star.createpdf.util.PDFUtils;
 import swati4star.createpdf.util.RealPathUtil;
+import swati4star.createpdf.util.SplitPDFUtils;
 import swati4star.createpdf.util.ViewFilesDividerItemDecoration;
 
 import static android.app.Activity.RESULT_OK;
@@ -57,7 +57,7 @@ public class SplitFilesFragment extends Fragment implements MergeFilesAdapter.On
     private String mPath;
     private MorphButtonUtility mMorphButtonUtility;
     private FileUtils mFileUtils;
-    private PDFUtils mPDFUtils;
+    private SplitPDFUtils mSplitPDFUtils;
     private BottomSheetUtils mBottomSheetUtils;
     private static final int INTENT_REQUEST_PICKFILE_CODE = 10;
 
@@ -127,7 +127,7 @@ public class SplitFilesFragment extends Fragment implements MergeFilesAdapter.On
     public void parse() {
         hideKeyboard(mActivity);
 
-        ArrayList<String> outputFilePaths = mPDFUtils.splitPDFByConfig(mPath,
+        ArrayList<String> outputFilePaths = mSplitPDFUtils.splitPDFByConfig(mPath,
                 mSplitConfitEditText.getText().toString());
         int numberOfPages = outputFilePaths.size();
         if (numberOfPages == 0) {
@@ -158,7 +158,7 @@ public class SplitFilesFragment extends Fragment implements MergeFilesAdapter.On
         mActivity = (Activity) context;
         mMorphButtonUtility = new MorphButtonUtility(mActivity);
         mFileUtils = new FileUtils(mActivity);
-        mPDFUtils = new PDFUtils(mActivity);
+        mSplitPDFUtils = new SplitPDFUtils(mActivity);
         mBottomSheetUtils = new BottomSheetUtils(mActivity);
     }
